Se organizan sevices y se agrega la consulta de participantes para el cierre de la etapa
