+++ src/main/java/com/mozcalti/gamingapp/jobs/EjecutaBatallaAutomaticamente.java
@@ -1,6 +1,6 @@
 package com.mozcalti.gamingapp.jobs;
 
-import com.mozcalti.gamingapp.service.BatallasService;
+import com.mozcalti.gamingapp.service.batallas.BatallasService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
