--- src/main/java/com/mozcalti/gamingapp/service/EtapasService.java
@@ -1,8 +1,0 @@
-package com.mozcalti.gamingapp.service;
-
-import com.mozcalti.gamingapp.commons.GenericServiceAPI;
-import com.mozcalti.gamingapp.model.Etapas;
-
-public interface EtapasService extends GenericServiceAPI<Etapas, Integer> {
-
-}
