+++ src/main/java/com/mozcalti/gamingapp/service/batallas/impl/BatallaParticipantesServiceImpl.java
@@ -1,9 +1,9 @@
-package com.mozcalti.gamingapp.service.impl;
+package com.mozcalti.gamingapp.service.batallas.impl;
 
 import com.mozcalti.gamingapp.commons.GenericServiceImpl;
 import com.mozcalti.gamingapp.model.BatallaParticipantes;
 import com.mozcalti.gamingapp.repository.BatallaParticipantesRepository;
-import com.mozcalti.gamingapp.service.BatallaParticipantesService;
+import com.mozcalti.gamingapp.service.batallas.BatallaParticipantesService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
