+++ src/main/java/com/mozcalti/gamingapp/service/equipo/EtapaEquipoService.java
@@ -1,4 +1,4 @@
-package com.mozcalti.gamingapp.service;
+package com.mozcalti.gamingapp.service.equipo;
 
 import com.mozcalti.gamingapp.commons.GenericServiceAPI;
 import com.mozcalti.gamingapp.model.EtapaEquipo;
