+++ src/main/java/com/mozcalti/gamingapp/service/equipo/impl/EquiposServiceImpl.java
@@ -1,9 +1,9 @@
-package com.mozcalti.gamingapp.service.impl;
+package com.mozcalti.gamingapp.service.equipo.impl;
 
 import com.mozcalti.gamingapp.commons.GenericServiceImpl;
 import com.mozcalti.gamingapp.model.Equipos;
 import com.mozcalti.gamingapp.repository.EquiposRepository;
-import com.mozcalti.gamingapp.service.EquiposService;
+import com.mozcalti.gamingapp.service.equipo.EquiposService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
