+++ src/main/java/com/mozcalti/gamingapp/service/torneo/TorneosService.java
@@ -1,4 +1,4 @@
-package com.mozcalti.gamingapp.service;
+package com.mozcalti.gamingapp.service.torneo;
 
 import com.mozcalti.gamingapp.commons.GenericServiceAPI;
 import com.mozcalti.gamingapp.exceptions.ValidacionException;
