+++ src/main/resources/application.yml
@@ -52,8 +52,8 @@ server:
       enabled: false
 
 cron:
-  mail-batallas: ${CRON_MAIL_BATTALLAS:0 */5 * * * *}
-  resultados-batallas: ${CRON_RESULTADOS_BATTALLAS:0 */5 * * * *}
+  mail-batallas: ${CRON_MAIL_BATTALLAS:0 0 1 * * *}
+  resultados-batallas: ${CRON_RESULTADOS_BATTALLAS:0 */1 * * * *}
   ejecuta-batallas: ${CRON_EJECUTA_BATTALLAS:0 */1 * * * *}
 
 mail:
